@@ -1,20 +1,18 @@
 module github.com/slarwise/yamlls
 
-go 1.21.5
+go 1.23.0
+
+toolchain go1.24.3
 
 require (
-	github.com/bmatcuk/doublestar/v4 v4.6.1
-	github.com/tidwall/gjson v1.17.1
+	github.com/goccy/go-yaml v1.18.0
 	github.com/xeipuuv/gojsonschema v1.2.0
 	go.lsp.dev/protocol v0.12.0
-	go.lsp.dev/uri v0.3.0
 )
 
 require (
 	github.com/segmentio/asm v1.2.0 // indirect
-	github.com/segmentio/encoding v0.4.0 // indirect
-	github.com/tidwall/match v1.1.1 // indirect
-	github.com/tidwall/pretty v1.2.1 // indirect
+	github.com/segmentio/encoding v0.4.1 // indirect
 	github.com/xeipuuv/gojsonpointer v0.0.0-20190905194746-02993c407bfb // indirect
 	github.com/xeipuuv/gojsonreference v0.0.0-20180127040603-bd5ef7bd5415 // indirect
 	go.lsp.dev/jsonrpc2 v0.10.0 // indirect
@@ -24,10 +22,7 @@ require (
 )
 
 require (
-	github.com/fatih/color v1.16.0 // indirect
-	github.com/goccy/go-yaml v1.11.3
-	github.com/mattn/go-colorable v0.1.13 // indirect
-	github.com/mattn/go-isatty v0.0.20 // indirect
-	golang.org/x/sys v0.19.0 // indirect
-	golang.org/x/xerrors v0.0.0-20231012003039-104605ab7028 // indirect
+	go.lsp.dev/uri v0.3.0 // indirect
+	golang.org/x/sys v0.33.0 // indirect
+	golang.org/x/xerrors v0.0.0-20240903120638-7835f813f4da // indirect
 )
